Move EmailQueryParams into dataclasses
