+++ lib/openstack_api/openstack_floating_ip.py
@@ -10,8 +10,8 @@ from openstack_api.openstack_connection import OpenstackConnection
 from openstack_api.openstack_identity import OpenstackIdentity
 from openstack_api.openstack_query import OpenstackQuery
 from openstack_api.openstack_wrapper_base import OpenstackWrapperBase
+from openstack_api.dataclasses import EmailQueryParams
 from structs.email_params import EmailParams
-from structs.email_query_params import EmailQueryParams
 from structs.smtp_account import SMTPAccount
